Move all models to a separate folder in `lib`
