+++ app/api/auth.py
@@ -1,33 +1,14 @@
-from typing import Optional
-
 from fastapi import APIRouter, HTTPException, status
-from pydantic import BaseModel
 
 from app.lib.auth.prisma import (
     encryptPassword,
     signJWT,
     validatePassword,
 )
+from app.lib.models.auth import SignIn, SignInOut, SignUp
 from app.lib.prisma import prisma
-from prisma.models import User
 
 router = APIRouter()
-
-
-class SignIn(BaseModel):
-    email: str
-    password: str
-
-
-class SignUp(BaseModel):
-    email: str
-    password: str
-    name: Optional[str] = None
-
-
-class SignInOut(BaseModel):
-    token: str
-    user: User
 
 
 @router.post("/auth/sign-in", tags=["auth"])
