+++ app/lib/models/agents.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Agent(BaseModel):
+    name: str
